pyproject.toml
[project]
name = "fedorentos-pet-feeder"
version = "0.1.0"
description = "Automated pet feeder with image recognition for 5 cats"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "viam-sdk>=0.20.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
